requirements.txt
altgraph==0.17.4
asgiref==3.7.2
blinker==1.7.0
boto3==1.34.61
botocore==1.34.61
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==3.13.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Flask==3.0.2
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
psycopg2-binary==2.9.9
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.31.0
s3transfer==0.10.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.0.7
Werkzeug==3.0.1
whitenoise==6.6.0
